airflow/requirements.txt
kaggle
google-cloud-storage
python-dotenv
# os
pinecone-client